Configure "jujutsu" to use "ediff" differently
Nix's Home Manager can configure Emac's "ediff" program as a merge
tool automatically, but when it does so, it has the "jj" tool invoke
"emacsclient" with its "-c" or "--create-frame" flag active, such that
"emacsclient" has Emacs create a new frame—giving it focus—and then
blocks until the session it started completes. This makes it difficult
to work with other existing Emacs windows, and behaves poorly when one
quits "ediff" without saving its output deliberately, wishing to abort
the attempted resolution.

Instead, define an alternative but similar merge tool, invoking
"ediff" without creating a fresh frame, operating asynchronously,
allowing the session to proceed without "ediff" blocking in wait of
its conclusion. Use the "fswatch" tool to block waiting on deletion of
a sentinel file that indicates that the "ediff" session is complete.