Centralize the "unfree" Nix package configuration
Configure the "mkDarwin" Nix function to accept an optional "pkgs"
parameter, allowing both nix-darwin and Home Manager to share the same
nixpkgs instance with its "unfree" package configuration. Doing so
eliminates the prior duplication of the "allowUnfreePredicate"
function in multiple files.

Include the "terraform" package newly in the allowed "unfree" set, as
its license changed recently such that the Nix project no longer
considers it to be "free".