Move Zoom application into the "desktop" profile
The "zoom-us" Nix package is not available on some platforms that are
newly desirable to support—such as "aarch64-linux"—so it should not be
in the "essential" profile, as that one should work on all
platforms. Instead, move "zoom-us" to the "desktop" profile where it
better belongs, and use the "lib.meta.availableOn" function to include
it conditionally only on supported platforms.